Project Structure & Emotes
Change project structure & Allow for emotes to be used in whole message
