+++ src/replacements.lua
@@ -1,0 +1,3 @@
+return {
+    {':sippy:', '<:sippy:1010770471500910693>'}
+}
